fix bug

+++ test/ferret.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int port, party;
 const static int threads = 1;
 
-void test_ferret(int party, NetIO *ios[threads], int num_ot) {
+void test_ferret(int party, NetIO *ios[threads], size_t num_ot) {
 	auto start = clock_start();
 	FerretCOT<NetIO> * ferretcot = new FerretCOT<NetIO>(party, threads, ios, true);
 	double timeused = time_from(start);
@@ -13,7 +13,7 @@ void test_ferret(int party, NetIO *ios[threads], int num_ot) {
 
 	// RCOT
 	// The RCOTs will be generated at internal memory, and copied to user buffer
-	int num = 1 << num_ot;
+	size_t num = 1 << num_ot;
 	cout <<"Active FERRET RCOT\t"<<double(num)/test_rcot<FerretCOT<NetIO>>(ferretcot, ios[0], party, num, false)*1e6<<" OTps"<<endl;
 
 	// RCOT inplace
@@ -30,7 +30,7 @@ int main(int argc, char** argv) {
 	for(int i = 0; i < threads; ++i)
 		ios[i] = new NetIO(party == ALICE?nullptr:"127.0.0.1",port+i);
 
-	int length = 20;
+	size_t length = 20;
 	if (argc > 3)
 		length = atoi(argv[3]);
 	test_ferret(party, ios, length);
